+++ index.html
@@ -10,7 +10,7 @@
         <link rel="stylesheet" href="headers.css">
         <link rel="stylesheet" href="main.css">
     </head>
-    <body id="background" class="gradient">
+    <body id="background" class="">
         <audio src="" autoplay>
             Your browser does not support playing audio files
         </audio>
@@ -22,9 +22,9 @@
             <span id="usScore">USA : 0</span>
             <span id="ussrScore">USSR : 0</span>
         </h2>
-        <button onclick="resetGame()" class="hidden">New Game</button>
+        <button onclick="" class="hidden">New Game</button>
         <div id="board"></div>
-        <script src="main.js"></script>
+        <script src="./javascript/main.js" type="module"></script>
     </body>
     <footer>
